// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  CUSTOMER: 'CUSTOMER',
  TECHNICIAN: 'TECHNICIAN',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  BLOCKED: 'BLOCKED',
  SUSPENDED: 'SUSPENDED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const Severity = {
  SMALL: 'SMALL',
  MEDIUM: 'MEDIUM',
  LARGE: 'LARGE'
} as const

export type Severity = (typeof Severity)[keyof typeof Severity]


export const VerificationStatus = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED'
} as const

export type VerificationStatus = (typeof VerificationStatus)[keyof typeof VerificationStatus]


export const RequestType = {
  AI_ESTIMATION: 'AI_ESTIMATION',
  CONSULTATION: 'CONSULTATION'
} as const

export type RequestType = (typeof RequestType)[keyof typeof RequestType]


export const RequestStatus = {
  PENDING: 'PENDING',
  WAITING_FOR_TECHNICIAN: 'WAITING_FOR_TECHNICIAN',
  TECHNICIAN_SELECTED: 'TECHNICIAN_SELECTED',
  ON_THE_WAY: 'ON_THE_WAY',
  ARRIVED: 'ARRIVED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus]


export const OfferStatus = {
  PENDING: 'PENDING',
  SUBMITTED: 'SUBMITTED',
  DECLINED: 'DECLINED',
  SELECTED: 'SELECTED',
  NOT_SELECTED: 'NOT_SELECTED'
} as const

export type OfferStatus = (typeof OfferStatus)[keyof typeof OfferStatus]


export const PointsTransactionType = {
  TOPUP: 'TOPUP',
  SPEND: 'SPEND',
  REFUND: 'REFUND',
  ADMIN_GRANT: 'ADMIN_GRANT'
} as const

export type PointsTransactionType = (typeof PointsTransactionType)[keyof typeof PointsTransactionType]